refactor(gkplus): make bulk-create internals private
RankData is mutable bookkeeping used only inside bulk_create.py; rename it to _RankData to mark it as such. Drop the backward-compatibility re-export block in g_k_plus_base.py (RankData, _klist_to_tree, _tree_to_klist, bulk_create_gkplus_tree) so the module surface matches its purpose. Callers that legitimately need these symbols (zip/conversion mixins, white-box tests of the bulk module, benchmarks) now import them directly from gplus_trees.g_k_plus.bulk_create.
